SendGasSettingsView refactored
